+++ Runtime/Events/GameEventFloat.cs
@@ -1,0 +1,7 @@
+using UnityEngine;
+
+namespace ProtoCasual.Core.Events
+{
+    [CreateAssetMenu(menuName = "ProtoCasual/Events/Float Event")]
+    public class GameEventFloat : GameEvent<float> { }
+}