+++ Runtime/Events/GameEventInt.cs
@@ -1,27 +1,7 @@
 using UnityEngine;
-using System.Collections.Generic;
-using System;
 
 namespace ProtoCasual.Core.Events
 {
-    [CreateAssetMenu(fileName = "New Int Event", menuName = "Game/Events/Int Event")]
-    public class GameEventInt : ScriptableObject
-    {
-        private event Action<int> OnEventRaised;
-
-        public void Raise(int value)
-        {
-            OnEventRaised?.Invoke(value);
-        }
-
-        public void RegisterListener(Action<int> listener)
-        {
-            OnEventRaised += listener;
-        }
-
-        public void UnregisterListener(Action<int> listener)
-        {
-            OnEventRaised -= listener;
-        }
-    }
+    [CreateAssetMenu(menuName = "ProtoCasual/Events/Int Event")]
+    public class GameEventInt : GameEvent<int> { }
 }