+++ Runtime/Events/GameEventString.cs
@@ -1,0 +1,7 @@
+using UnityEngine;
+
+namespace ProtoCasual.Core.Events
+{
+    [CreateAssetMenu(menuName = "ProtoCasual/Events/String Event")]
+    public class GameEventString : GameEvent<string> { }
+}